# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vesta Family Logistics OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inbox:
    get:
      operationId: listInboxItems
      tags: [inbox]
      summary: List inbox items
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [received, classifying, ready_for_review, approved, dismissed, failed, manual_review]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of inbox items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboxItem"
    post:
      operationId: createInboxItem
      tags: [inbox]
      summary: Create a new inbox item (manual in-app input)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboxItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
  /inbox/{id}:
    get:
      operationId: getInboxItem
      tags: [inbox]
      summary: Get a single inbox item with its suggested actions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItemDetail"
        "404":
          description: Not found
  /inbox/{id}/classify:
    post:
      operationId: classifyInboxItem
      tags: [inbox]
      summary: Trigger AI classification of an inbox item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Classification triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"

  # ── Suggested Actions ───────────────────────────────────────────
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List smart rules
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a smart rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /rules/{id}:
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /rules/{id}/toggle:
    post:
      operationId: toggleRule
      tags: [rules]
      summary: Pause or resume a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"

  # ── Household ────────────────────────────────────────────────────
...
  /patterns:
    get:
      operationId: listPatterns
      tags: [patterns]
      summary: List pattern observations and suggestions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [accumulating, threshold_met, suggested, rule_created, dismissed]
      responses:
        "200":
          description: Pattern observations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatternObservation"
  /patterns/{id}/accept:
    post:
      operationId: acceptPattern
      tags: [patterns]
      summary: Accept a pattern suggestion (creates a rule)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted — rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternObservation"
  /patterns/{id}/dismiss:
    post:
      operationId: dismissPattern
      tags: [patterns]
      summary: Dismiss a pattern suggestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternObservation"
components:
  schemas:
...
    InboxItem:
      type: object
      required: [id, source, raw_content, status, created_at]
      properties:
        id:
          type: integer
        source:
          type: string
          enum: [whatsapp, email, manual, photo]
        raw_content:
          type: string
        media_url:
          type: ["string", "null"]
        status:
          type: string
          enum: [received, classifying, ready_for_review, approved, dismissed, failed, manual_review]
        sender_name:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        actions_count:
          type: integer
    InboxItemDetail:
      type: object
      required: [id, source, raw_content, status, created_at, actions]
      properties:
        id:
          type: integer
        source:
          type: string
        raw_content:
          type: string
        media_url:
          type: ["string", "null"]
        status:
          type: string
        sender_name:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        actions:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedAction"
    InboxItemInput:
      type: object
      required: [raw_content, source]
      properties:
        raw_content:
          type: string
        source:
          type: string
          enum: [manual, whatsapp, email, photo]
        sender_name:
          type: string

    # ── Suggested Actions ────────────────────────────────────────
    SuggestedAction:
      type: object
      required: [id, inbox_item_id, category, type, title, approval_level, status, confidence]
      properties:
        id:
          type: integer
        inbox_item_id:
          type: integer
        category:
          type: string
          enum: [escola, saude, casa, social, logistica, refeicoes, servicos, outros]
        type:
          type: string
          enum: [event, task, reminder, fyi]
        title:
          type: string
        datetime:
          type: ["string", "null"]
        suggested_owner:
          type: ["string", "null"]
        approval_level:
          type: string
          enum: [soft, one_tap, explicit]
        confidence:
          type: number
        status:
          type: string
          enum: [pending, approved, dismissed, auto_handled]
        notes:
          type: ["string", "null"]
        cascade_check_needed:
          type: boolean
        workflow_tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    Rule:
      type: object
      required: [id, name, category, trigger_desc, action_desc, approval_level, confidence, active]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [escola, saude, casa, social, logistica, refeicoes, servicos, outros]
        trigger_desc:
          type: string
        action_desc:
          type: string
        approval_level:
          type: string
          enum: [soft, one_tap, explicit]
        confidence:
          type: number
        active:
          type: boolean
        origin:
          type: string
          enum: [system_template, user_created, pattern_suggested]
        times_triggered:
          type: integer
        times_approved:
          type: integer
        times_dismissed:
          type: integer
        created_at:
          type: string
          format: date-time
    RuleInput:
      type: object
      required: [name, category, trigger_desc, action_desc, approval_level]
      properties:
        name:
          type: string
        category:
          type: string
        trigger_desc:
          type: string
        action_desc:
          type: string
        approval_level:
          type: string
          enum: [soft, one_tap, explicit]
    RuleUpdate:
      type: object
      properties:
        name:
          type: string
        trigger_desc:
          type: string
        action_desc:
          type: string
        approval_level:
          type: string
          enum: [soft, one_tap, explicit]
        active:
          type: boolean

    # ── Patterns ──────────────────────────────────────────────────
    PatternObservation:
      type: object
      required: [id, type, description, occurrences, confidence, status]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [sender, sequence, temporal, ownership, absence, seasonal]
        description:
          type: string
        occurrences:
          type: integer
        confidence:
          type: number
        status:
          type: string
          enum: [accumulating, threshold_met, suggested, rule_created, dismissed]
        evidence:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time

    # ── Dashboard ─────────────────────────────────────────────────
...