# lib/api-zod/src/generated/types/inboxItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vesta Family Logistics OS API
 * OpenAPI spec version: 0.1.0
 */
import type { InboxItemSource } from "./inboxItemSource";
import type { InboxItemStatus } from "./inboxItemStatus";

export interface InboxItem {
  id: number;
  source: InboxItemSource;
  raw_content: string;
  /** @nullable */
  media_url?: string | null;
  status: InboxItemStatus;
  /** @nullable */
  sender_name?: string | null;
  created_at: Date;
  actions_count?: number;
}